requirements.txt
argon2_cffi==18.3.0
psycopg2==2.7.5
psycopg2-binary==2.7.5

# ** Django **
Django==2.2.5
django-background-tasks==1.2.0
django-ckeditor==5.7.1
django-crontab==0.7.1
django-extensions==2.2.1

# ** quality **
coverage==4.5.4
flake8==3.7.8  # (pep8 + pyflakes)
pydocstyle==4.0.1
pylint==2.4.2
pylint-django==2.0.11

# ** testing **
django-dynamic-fixture==2.0.0
django-webtest==1.9.7